Extracted all /practice stuff into seperate directory & more restructuring
