+++ src/commands/keywords.py
@@ -1,6 +1,6 @@
 from bot import bot, discord, pymongo, keywords, time
-from constants import LINK, GUILD_ID, MODLOG_CHANNEL_ID
-from mongodb import kwdb
+from utils.constants import LINK, GUILD_ID, MODLOG_CHANNEL_ID
+from utils.mongodb import kwdb
 from utils.roles import is_moderator, is_bot_developer
