+++ src/commands/moderation.py
@@ -2,8 +2,8 @@ from bot import bot, discord, pymongo, datetime, time
 from commands.dms import send_dm
 from utils.bans import is_banned
 from utils.roles import is_chat_moderator, is_moderator
-from mongodb import gpdb, punishdb
-from constants import GUILD_ID
+from utils.mongodb import gpdb, punishdb
+from utils.constants import GUILD_ID
 
 
 def convert_time(time: tuple[str, str, str, str]) -> str:
