+++ src/commands/practice/__init__.py
@@ -1,0 +1,3 @@
+from .practice import (
+    close_session,
+)
