+++ src/commands/practice/ui/DisabledButtonsView.py
