+++ src/commands/practice/ui/JoinSessionListView.py
