+++ src/events/on_member_join.py
@@ -1,6 +1,6 @@
 from bot import bot, discord
 from commands.dms import send_dm
-from constants import GUILD_ID, WELCOME_CHANNEL_ID, BETA
+from utils.constants import GUILD_ID, WELCOME_CHANNEL_ID, BETA
 
 welcome_embed = discord.Embed.from_dict(
     {
