+++ src/events/on_raw_reaction_add.py
@@ -1,6 +1,6 @@
 from bot import discord, bot, requests
-from constants import GUILD_ID
-from mongodb import gpdb, rrdb
+from utils.constants import GUILD_ID
+from utils.mongodb import gpdb, rrdb
 from utils.roles import is_moderator
