+++ src/events/on_voice_state_update.py
@@ -1,5 +1,5 @@
 from bot import bot, discord
-from constants import GUILD_ID
+from utils.constants import GUILD_ID
 
 
 @bot.event
